WEBAPP-208: Move files back to layout
