+++ src/modules/layout/components/Footer.css
@@ -1,6 +1,6 @@
-@import url('../../../app/constants/helpers.css');
-@import url('../../../app/constants/colors.css');
-@import url('../../../app/constants/dimensions.css');
+@import url('../../app/constants/helpers.css');
+@import url('../../app/constants/colors.css');
+@import url('../../app/constants/dimensions.css');
 
 .footer {
   display: flex;
