+++ src/modules/layout/components/HeaderActionBar.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Link from 'redux-first-router-link'
 
 import { ActionItems } from './HeaderActionBar.styles'
-import HeaderActionItem from '../../HeaderActionItem'
+import HeaderActionItem from '../HeaderActionItem'
 
 /**
  * Designed to work with Header. In the ActionBar you can display icons as link or dropDown involving actions like
