+++ src/modules/layout/components/HeaderNavigationBar.js
