+++ src/modules/layout/components/__tests__/HeaderNavigationBar.spec.js
