+++ src/modules/layout/components/__tests__/Toolbar.spec.js
