+++ src/modules/layout/components/__tests__/ToolbarAnchor.spec.js
