--- src/modules/layout/components/header/HeaderNavigationBar.css
@@ -1,13 +1,0 @@
-@import url('../../../app/constants/dimensions.css');
-@import url('../../../app/constants/helpers.css');
-@import url('../../../app/constants/fonts.css');
-
-.navigation-bar {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.navigation-bar.hidden {
-  display: none;
-}
