+++ src/modules/layout/containers/LocationHeader.js
@@ -8,8 +8,8 @@ import compose from 'lodash/fp/compose'
 import LanguageSelector from '../../common/containers/LanguageSelector'
 import searchIcon from '../assets/magnifier.svg'
 import landingIcon from '../assets/location-icon.svg'
-import Header from 'modules/layout/components/header/Header'
-import HeaderNavigationItem from '../components/header/HeaderNavigationItem'
+import Header from 'modules/layout/components/Header'
+import HeaderNavigationItem from '../components/HeaderNavigationItem'
 import HeaderActionItem from '../HeaderActionItem'
 import { EXTRAS_ROUTE, goToExtras } from '../../app/routes/extras'
 import { CATEGORIES_ROUTE, goToCategories } from '../../app/routes/categories'
