+++ src/modules/layout/containers/LocationLayout.js
@@ -5,11 +5,11 @@ import { connect } from 'react-redux'
 
 import CityModel from 'modules/endpoint/models/CityModel'
 
-import GeneralHeader from '../components/header/GeneralHeader'
+import GeneralHeader from '../components/GeneralHeader'
 import Layout from '../components/Layout'
-import GeneralFooter from '../components/footer/GeneralFooter'
+import GeneralFooter from '../components/GeneralFooter'
 import LocationHeader from './LocationHeader'
-import LocationFooter from '../components/footer/LocationFooter'
+import LocationFooter from '../components/LocationFooter'
 import { CATEGORIES_ROUTE } from '../../app/routes/categories'
 import { EVENTS_ROUTE } from '../../app/routes/events'
 import { EXTRAS_ROUTE } from '../../app/routes/extras'
